Almost done
This is kinda the "release", the program is fully functional but is missing a countdown timer. It will now restart every (your input) hour. It will save your information in a restartsettings.txt file. Where it saves what restart times you have set and also the language you have chosen.